backend/requirements.txt
django
djangorestframework
django-cors-headers
pandas
numpy
scikit-learn
joblib

